+++ node/src/main/kotlin/net/corda/node/services/transactions/RaftValidatingNotaryService.kt
@@ -3,7 +3,6 @@ package net.corda.node.services.transactions
 import net.corda.core.crypto.Party
 import net.corda.core.flows.FlowLogic
 import net.corda.core.node.services.TimestampChecker
-import net.corda.flows.ValidatingNotaryFlow
 
 /** A validating notary service operated by a group of mutually trusting parties, uses the Raft algorithm to achieve consensus. */
 class RaftValidatingNotaryService(val timestampChecker: TimestampChecker,